+++ crates/app-gpui/ui/plugin.rs
@@ -216,7 +216,6 @@ impl PlayerView {
     quick_add_plugin_handler!(quick_add_transient_shaper, QuickAddTransientShaper, sotf_audio_player::PluginType::TransientShaper);
     quick_add_plugin_handler!(quick_add_saturation, QuickAddSaturation, sotf_audio_player::PluginType::Saturation);
     quick_add_plugin_handler!(quick_add_dynamic_eq, QuickAddDynamicEq, sotf_audio_player::PluginType::DynamicEq);
-    quick_add_plugin_handler!(quick_add_fir_designer, QuickAddFirDesigner, sotf_audio_player::PluginType::FirDesigner);
     quick_add_plugin_handler!(quick_add_linear_phase_eq, QuickAddLinearPhaseEq, sotf_audio_player::PluginType::LinearPhaseEq);
     quick_add_plugin_handler!(quick_add_spectral_compressor, QuickAddSpectralCompressor, sotf_audio_player::PluginType::SpectralCompressor);